Add ghost table view controller to display placeholder view
